w3c-reports/12092014/efl/runned-tests/html-media-capture/idlharness-expected.txt

CONSOLE MESSAGE: Deprecated attempt to access property 'capture' on a non-HTMLInputElement object.
HTML Media Capture IDL tests


FAIL HTMLInputElement interface: attribute capture assert_throws: getting property on prototype object must throw TypeError function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1270:26
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
http://localhost:8800/html-media-capture/idlharness.html:46:23
global code@http://localhost:8800/html-media-capture/idlharness.html:47:9)
PASS HTMLInputElement must be primary interface of [object HTMLInputElement] 
PASS Stringification of [object HTMLInputElement] 
PASS HTMLInputElement interface: [object HTMLInputElement] must inherit property "capture" with the proper type (0) 
